Implement entry point overwriting and writing shellcode into binary.
While elfector does not support pie binaries yet, this patch shows
a PoC with non pie binaries, and a minimal shellcode that only jumps
to the original entry point with harcoded addresses.
It is possible to see with gdb the shellcode being executed as a
proof for the PoC working.