Use RIP-relative addressing in shellcode to support PIE binaries
Now using 0xdeadbeef as a place-holder value in the shellcode and
inside make_shellcode, dynamically patching the shellcode with the
correct offset from the original entry point to the new entry point.
As of now, the deadbeef offset in the shellcode is hard-coded but in
the future should not be.